Augment data with information from a survival_ln_mixture object

Include information about hazard and survival distribution for each individual in a dataset.

Usage

# S3 method for class 'survival_ln_mixture'
augment(x, newdata, eval_time, ...)

Arguments

x

A survival_ln_mixture object.

newdata

A base::data.frame() or tibble::tiblle() containing all the original predictors used to create x.

eval_time

a vector with the times where the hazard and survival distribuition will be evaluated.

...

Not used.

Value

A tibble::tibble() with the original covariates and ther survvial and hazard distributions.