Tidying method for a Lognormal Mixture model.

These method tidy the estimates from survival_ln_mixture fits into a summary.

Usage

# S3 method for survival_ln_mixture
tidy(
  x,
  effects = "fixed",
  conf.int = FALSE,
  conf.level = 0.9,
  digits = NULL,
  ...
)

Arguments

x

Fitted model object.

effects

A character vector including one or more of "fixed" and "auxiliary.

conf.int

If TRUE columns for lower (conf.low) and upper (conf.high) bounds of the posterior uncertainty intervals are included.

conf.level

A number between 0 and 1 indicating the desired probability mass to include in the intervals. Only used if conf.int = TRUE.

digits

How many significant digits should be displayed?

...

Not used.

Value

A data.frame without rownames. When effects="fixed" (the default), tidy.survival_ln_mixutre returns one row for each coefficient for each component of the mixture with three columns:

term

The name of the corresponding term in the model.

estimate

A point estimate of the coefficient (posterior median).

std.error

A standard error for the point estimate based on mad. See the Uncertainty estimates section in print.stanreg for more details.

Setting effects="auxiliary" will select the remaining parameters:

phi_a

Dispersion parameter for the first componente of the mixture.

phi_b

Dispersion parameter for the second componente of the mixture.

theta_a

The weigth of the first component of the mixture.

Examples

require(survival)
lung$sex <- factor(lung$sex)
set.seed(1)
mod2 <- survival_ln_mixture(Surv(time, status == 2) ~ sex, lung, intercept = TRUE)
#> Skipping EM Algorithm
tidy(mod2)
#> # A tibble: 4 × 3
#>   term          estimate std.error
#>   <chr>            <dbl>     <dbl>
#> 1 (Intercept)_1    5.58     0.0735
#> 2 sex2_1           0.421    0.133 
#> 3 (Intercept)_2    2.50     0.0510
#> 4 sex2_2          -0.891    0.121 
tidy(mod2, conf.int = TRUE)
#> # A tibble: 4 × 5
#>   term          estimate std.error conf.low conf.high
#>   <chr>            <dbl>     <dbl>    <dbl>     <dbl>
#> 1 (Intercept)_1    5.58     0.0735    5.48      5.67 
#> 2 sex2_1           0.421    0.133     0.262     0.586
#> 3 (Intercept)_2    2.50     0.0510    2.43      2.57 
#> 4 sex2_2          -0.891    0.121    -1.07     -0.735
tidy(mod2, effects = c("fixed", "auxiliary"), conf.int = TRUE)
#> # A tibble: 7 × 5
#>   term          estimate std.error conf.low conf.high
#>   <chr>            <dbl>     <dbl>    <dbl>     <dbl>
#> 1 (Intercept)_1    5.58     0.0735    5.48      5.67 
#> 2 sex2_1           0.421    0.133     0.262     0.586
#> 3 (Intercept)_2    2.50     0.0510    2.43      2.57 
#> 4 sex2_2          -0.891    0.121    -1.07     -0.735
#> 5 phi_1            1.40     0.157     1.20      1.61 
#> 6 phi_2           66.2     39.9      25.5     148.   
#> 7 eta_1            0.965    0.0120    0.949     0.978